separate ui and api tasks upd build.gradle
