+++ src/test/java/api/Spec.java
@@ -1,0 +1,4 @@
+package api;
+
+public class Spec {
+}
