+++ src/test/java/api/responses/AddToFavouriteResponeModel.java
@@ -1,0 +1,9 @@
+package api.responses;
+
+import lombok.Data;
+
+@Data
+public class AddToFavouriteResponeModel {
+    String result;
+
+}
