+++ src/test/java/api/responses/DeleteFromFavouriteResponeModel.java
@@ -1,0 +1,9 @@
+package api.responses;
+
+import lombok.Data;
+
+@Data
+public class DeleteFromFavouriteResponeModel {
+    String result;
+
+}
