+++ src/test/java/tests/web/WebTests.java
@@ -10,7 +10,7 @@ import tests.TestBase;
 @Tag("web")
 @Owner("staya_kotyat")
 @Feature("Movies page")
-@DisplayName("Movies page")
+@DisplayName("Web tests")
 public class WebTests extends TestBase {
 
     @Test
